Create new native style crate
