--- azul/src/styles/native_linux.css
@@ -1,41 +1,0 @@
-* {
-    font-size: 12px;
-    font-family: sans-serif;
-    color: #4c4c4c;
-}
-
-.__azul-native-button {
-    border: 1px solid #b7b7b7;
-    border-radius: 4px;
-    box-shadow: 0px 0px 3px #c5c5c5ad;
-    background: linear-gradient(#fcfcfc, #efefef);
-    text-align: center;
-    flex-direction: column;
-    justify-content: center;
-}
-
-.__azul-native-label {
-    text-align: center;
-    flex-direction: column;
-    justify-content: center;
-}
-
-.__azul-native-input-text {
-    background-color: white;
-    height: 14px;
-    border: 1px solid #9b9b9b;
-    padding: 1px;
-    overflow: hidden;
-    text-align: left;
-    flex-direction: row;
-    align-content: flex-end;
-    justify-content: flex-end;
-}
-
-.__azul-native-input-text:hover {
-    border: 1px solid #4286f4;
-}
-
-.__azul-native-input-text-label {
-
-}
